+++ movie-rental-system-backend/src/main/java/com/movie_rental_system/backend/service/EmployeeService.java
@@ -1,6 +1,8 @@
 package com.movie_rental_system.backend.service;
 
 import com.movie_rental_system.backend.entity.Employee;
+import com.movie_rental_system.backend.exception.EmployeeAlreadyExistsException;
+import com.movie_rental_system.backend.exception.EmployeeNotFoundException;
 import com.movie_rental_system.backend.repository.EmployeeRepository;
 import org.springframework.stereotype.Service;
 
@@ -21,17 +23,24 @@ public class EmployeeService {
 
     // find specific employee
     public Employee getEmployeeByName(String employeeName) {
-        return employeeRepository.findById(employeeName).get();
+        Employee employee = employeeRepository.findById(employeeName).orElse(null);
+        if (employee == null)
+            throw new EmployeeNotFoundException("Employee with name " + employeeName + " not found");
+        return employee;
     }
 
     // save employee
     public Employee addEmployee(Employee employee) {
+        if(employeeRepository.existsById(employee.getUser_name()))
+            throw new EmployeeAlreadyExistsException("Employee with name " + employee.getUser_name() + " already exists");
         return employeeRepository.save(employee);
     }
 
     // delete employee
     public Employee deleteEmployee(String employeeName) {
-        Employee employee = employeeRepository.findById(employeeName).get();
+        Employee employee = employeeRepository.findById(employeeName).orElse(null);
+        if (employee == null)
+            throw new EmployeeNotFoundException("Employee with name " + employeeName + " not found");
         employeeRepository.deleteById(employeeName);
         return employee;
     }
@@ -41,7 +50,7 @@ public class EmployeeService {
         if(employeeRepository.existsById(employeeName)) {
             return employeeRepository.save(employee);
         }
-        return null;
+        throw new EmployeeNotFoundException("Employee with name " + employeeName + " not found");
     }
